LayoutTests/http/tests/security/dataURL/xss-DENIED-from-javascript-url-window-open-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,%3Chtml%3E%3Cbody%3E%3Cp%20id='accessMe'%3EPASS:%20Cross%20frame%20access%20from%20a%20data:%20URL%20was%20denied.%3C/p%3E%3Cp%3EInner-inner%20iframe.%3C/p%3E%3Ciframe%3E%3C/iframe%3E%3C/body%3E%3C/html%3E from frame with URL data:text/html,%3Chtml%3E%3Cbody%3E%3Ciframe%20src=%22data:text/html,%3Chtml%3E%3Cbody%3E%3Cp%20id='accessMe'%3EPASS:%20Cross%20frame%20access%20from%20a%20data:%20URL%20was%20denied.%3C/p%3E%3Cp%3EInner-inner%20iframe.%3C/p%3E%3Ciframe%3E%3C/iframe%3E%3C/body%3E%3C/html%3E%22%3E%3C/iframe%3E%3Cp%3EInner%20iframe.%3C/p%3E%3Cscript%3Evar%20innerFrame%20=%20frames[0];var%20testDone%20=%20false;setTimeout(test,%201);setTimeout(function()%20{if%20(!testDone)%20{alert('FAIL:%20Test%20timed%20out');}},%202000);function%20test()%20{var%20flag%20=%20innerFrame[0];if%20(!flag)%20{setTimeout(test,%201);return;}try%20{if%20(innerFrame.document.body%20&&%20innerFrame.document.getElementById('accessMe'))%20{innerFrame.document.getElementById('accessMe').innerHTML%20=%20'FAIL:%20Cross%20frame%20access%20from%20a%20data:%20URL%20was%20allowed.';testDone%20=%20true;return;}}%20catch%20(e)%20{}}%3C/script%3E%3C/body%3E%3C/html%3E. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-from-javascript-url-window-open.html from frame with URL data:text/html,%3Chtml%3E%3Cbody%3E%3Cp%3EOpened%20Frame.%3C/p%3E%3Cscript%3Eopener.document.getElementById('accessMe').innerHTML%20=%20'FAIL:%20Access%20from%20a%20window%20opened%20with%20a%20data:%20URL%20was%20allowed!';%3C/script%3E%3Ciframe%3E%3C/iframe%3E%3C/body%3E%3C/html%3E. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: TypeError: Undefined value
Opener Frame

PASS: Access from a window opened with a data: URL was denied.
